adding some more tests
