+++ Tests/UnitTests/Core/Detector/CylindricalContainerBuilderTests.cpp
@@ -26,6 +26,7 @@
 #include "Acts/Surfaces/DiscSurface.hpp"
 #include "Acts/Surfaces/RadialBounds.hpp"
 #include "Acts/Surfaces/Surface.hpp"
+#include "Acts/Tests/CommonHelpers/CylindricalDetector.hpp"
 #include "Acts/Utilities/BinningType.hpp"
 #include "Acts/Utilities/Enumerate.hpp"
 #include "Acts/Utilities/Logger.hpp"
@@ -40,60 +41,10 @@
 #include <vector>
 
 using namespace Acts;
+using namespace Acts::Test;
 using namespace Acts::Experimental;
 
 GeometryContext tContext;
-
-/// @brief A mockup volume builder, it generates volumes with
-/// a single surface filled in in order to use the CylindricalContainerBuilder
-/// infrastructure.
-template <typename surface_type, typename surface_bounds_type>
-class CylindricalVolumeBuilder : public IDetectorComponentBuilder {
- public:
-  CylindricalVolumeBuilder(const Transform3& transform,
-                           const CylinderVolumeBounds& vBounds,
-                           const surface_bounds_type& sBounds,
-                           const std::string& vName)
-      : IDetectorComponentBuilder(),
-        m_transform(transform),
-        m_volumeBounds(vBounds),
-        m_surfaceBounds(sBounds),
-        m_name(vName) {}
-
-  DetectorComponent construct(
-      [[maybe_unused]] const GeometryContext& gctx) const final {
-    // The outgoing root volumes
-    std::vector<std::shared_ptr<DetectorVolume>> rootVolumes;
-
-    // Ingredients
-    auto surface = Surface::makeShared<surface_type>(
-        (m_transform), std::make_shared<surface_bounds_type>(m_surfaceBounds));
-
-    auto bounds = std::make_unique<CylinderVolumeBounds>(m_volumeBounds);
-    auto portalGenerator = defaultPortalGenerator();
-    auto volume = DetectorVolumeFactory::construct(
-        portalGenerator, tContext, m_name, m_transform, std::move(bounds),
-        {surface}, {}, tryNoVolumes(), tryAllPortalsAndSurfaces());
-
-    // Add to the roots
-    rootVolumes.push_back(volume);
-
-    DetectorComponent::PortalContainer dContainer;
-    for (auto [ip, p] : enumerate(volume->portalPtrs())) {
-      dContainer[ip] = p;
-    }
-    return DetectorComponent{
-        {volume},
-        dContainer,
-        RootDetectorVolumes{rootVolumes, tryRootVolumes()}};
-  }
-
- private:
-  Transform3 m_transform;
-  CylinderVolumeBounds m_volumeBounds;
-  surface_bounds_type m_surfaceBounds;
-  std::string m_name;
-};
 
 class VolumeGeoIdGenerator : public IGeometryIdGenerator {
  public:
