+++ Tests/UnitTests/Plugins/Detray/CMakeLists.txt
@@ -1,3 +1,4 @@
 set(unittest_extra_libraries ActsPluginDetray)
+add_unittest(DetrayConverter DetrayConverterTests.cpp)
 add_unittest(DetrayGeometryConverter DetrayGeometryConverterTests.cpp)
 add_unittest(DetrayMaterialConverter DetrayMaterialConverterTests.cpp)
